Refactor v2.0
