+++ cypress/integration/amazon/BuyNowNotAProduct.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="Cypress" />
 
-import { addProductNotA, buyNow } from "../../AmazonFunctions/amazon/Products";
+import { addProductNotA, buyNow } from "../../functions/amazon/Products";
 
 describe("Test Suite for not logging user buying a product not available", function () {
   it("[BuyNow] Try to buy a product not available", function () {
