+++ cypress/integration/amazon/CYP-14.spec.js
@@ -1,7 +1,7 @@
 // type definitions for Cypress object "cy"
 /// <reference types="cypress" />
 
-import { CYP14 } from "../../AmazonFunctions/amazon/CYP-14";
+import { CYP14 } from "../../functions/amazon/CYP-14";
 
 describe("CYP-14 add to cart without login increasing amount", function () {
   it("CYP-14 add to cart without login increasing amount", CYP14);
