+++ cypress/integration/amazon/CYP-2.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="Cypress" />
 
-import { loginNow, navBarCheck } from "../../AmazonFunctions/amazon/Navbar";
+import { loginNow, navBarCheck } from "../../functions/amazon/Navbar";
 
 describe("Test Suite for egiftCards", function () {
   it("[NavBar] Principal Menu", function () {
