+++ cypress/integration/amazon/CYP-24.spec.js
@@ -1,7 +1,7 @@
 // type definitions for Cypress object "cy"
 /// <reference types="cypress" />
 
-import { CYP24 } from "../../AmazonFunctions/amazon/CYP-24";
+import { CYP24 } from "../../functions/amazon/CYP-24";
 
 describe("CYP-24 forgot password", function () {
   it("CYP-24 forgot password", CYP24);
