+++ cypress/integration/amazon/CYP-26.spec.js
@@ -1,7 +1,7 @@
 // type definitions for Cypress object "cy"
 /// <reference types="cypress" />
 
-import { CYP26 } from "../../AmazonFunctions/amazon/CYP-26";
+import { CYP26 } from "../../functions/amazon/CYP-26";
 
 describe("CYP-26 Other issues with Sign-In", function () {
   it("CYP-26 Other issues with Sign-In", CYP26);
