+++ cypress/integration/youtube/CYP-59_360view.spec.js
@@ -1,7 +1,7 @@
 // type definitions for Cypress object "cy"
 /// <reference types="cypress" />
 
-import { CYP59 } from "../../AmazonFunctions/CYP-59";
+import { CYP59 } from "../../functions/youtube/CYP-59";
 
 describe("CYP-59 360 view in Youtube", function () {
   it("CYP-59 360 view in Youtube", CYP59);
